src/test/java/io/github/vlad324/n221/SolutionTest.java
package io.github.vlad324.n221;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {

    private final Solution solution = new Solution();

    private static Stream<Arguments> testDataProvider() {
        return Stream.<Arguments>builder()
            .add(arguments(new char[][]{
                {'1', '0', '1', '0', '0'},
                {'1', '0', '1', '1', '1'},
                {'1', '1', '1', '1', '1'},
                {'1', '0', '0', '1', '1'}
            }, 4))
            .add(arguments(new char[][]{
                {'1', '1', '1', '1', '1'},
                {'1', '1', '1', '1', '1'},
                {'1', '1', '1', '1', '1'},
                {'1', '1', '1', '1', '1'},
                {'1', '1', '1', '1', '1'}
            }, 25))
            .add(arguments(new char[][]{
                {'0', '1'},
                {'1', '0'}
            }, 1))
            .add(arguments(new char[][]{{'0'}}, 0))
            .build();
    }

    @ParameterizedTest
    @MethodSource("testDataProvider")
    void should_produce_expected_result(char[][] matrix, int expected) {
        // when
        final var actual = solution.maximalSquare(matrix);

        // then
        assertThat(actual).isEqualTo(expected);
    }

}